Update README.md
Added descriptions of constants in the PHP config file:

The ADOBE_BRAND is what you will be calling the service to your users. Some places rebrand as "Capitol Connect".
The other values are all SCOs (which are Shareable Content Objects and represent meetings, courses, and just about any content that can be created on Adobe Connect. https://helpx.adobe.com/adobe-connect/webservices/getting-started-connect-web-services.html) that are unique to your instance of Adobe Connect. I don't use Adobe Connect any more, but I think if you are looking a folder for meetings or users in the web portal, the sco number is visible in the links for each item. If you can't find it there, you will have to use the API to query the item for its SCO.
Essentially, there is one folder that houses all the meetings in your instance.  Use that folder SCO as ADOBE_MFOLD. Alternatively, if you just want to work with a subset of meetings, create a folder for them and use that folder SCO.Create a folder for deleted meeting and use its sco for ADOBE_DFOLDCreate a folder for unassigned meetings and use its sco for ADOBE_UFOLDCreate a user named 'breakout' with admin hosting permissions for all meetings and put its SCO in ADOBE_BRKADOBE_ADM holds all the SCOs of users who you want the program to automatically grant hosting/viewing privileges when it creates a meeting. The SCO numbers are separated by a space.